refactor : util 패키지 상위 계층으로 분리
