+++ src/main/java/roomescape/util/DateTimeConfiguration.java
@@ -1,4 +1,4 @@
-package roomescape.service.util;
+package roomescape.util;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
