+++ src/test/java/roomescape/util/OldDateTimeFormatter.java
@@ -2,7 +2,6 @@ package roomescape.util;
 
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
-import roomescape.service.util.DateTimeFormatter;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
